pogema_port/requirements.txt
pogema>=1.4.0
pogema-toolbox>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
colorama>=0.4.6
